Software License Changes
The Creative Commons Attribution-NonCommercial-NoDerivatives 4.0
International ("CC BY-ND 4.0") license originally used with the BNRBot
software and source code is not valid for software licensing. Since the
original license is not valid for software licensing, the copyright holder
is now upgrading the license from its original version.

As of January 20, 2017, the BNRBot software will be licensed under the GNU
General Public License v3 ("GPLv3").

All previously distributed copies and modifications of this software are
hereby licensed under the GPLv3 license retroactively.